vblankmanager: fix errors 3

+++ src/vblankmanager.cpp
@@ -225,7 +225,7 @@ void __attribute__((optimize("-fno-unsafe-math-optimizations") )) vblankThreadRu
 			
 			if ( diff_to_half > diff_to_full+1)
 			{
-				offset=g_uRollingMaxDrawTime=rollingMaxDrawTime=(centered_mean + std::max(std::abs(max_drawtime+avg_drawtime-static_cast<int64_t>(drawTime)), static_cast<int64_t>(0))*refresh/60)/4;
+				offset=g_uRollingMaxDrawTime=rollingMaxDrawTime=(centered_mean + std::max(std::abs(static_cast<int64_t>(max_drawtime)+avg_drawtime-static_cast<int64_t>(drawTime)), static_cast<int64_t>(0))*refresh/60)/4;
 				
 			}
 			else
